backend/requirements.txt
# Flask core
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23

# Security
Werkzeug==2.3.7
bcrypt==4.0.1

# Data processing (for analytics)
pandas==2.1.4
numpy==1.26.2

# Streamlit for interactive dashboards
streamlit==1.28.1
plotly==5.17.0
streamlit-plotly-events==0.0.6

# Development
python-dotenv==1.0.0

# Testing (optional)
pytest==7.4.3
pytest-flask==1.2.0

# Email (future use)
Flask-Mail==0.9.1

# File handling
python-magic==0.4.27
